Update: atomic design restructuring
