+++ src/components/molecules/index.js
@@ -1,2 +1,3 @@
 export * from './formItem';
 export * from './formField';
+export * from './movieTile';
