+++ src/components/organisms/form/form.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/colors.scss";
+@import "../../../styles/colors.scss";
 
 .submitButton {
   border-radius: 0.4rem;
