+++ src/components/organisms/publicHeader/publicHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/colors.scss';
+@import '../../../styles/colors.scss';
 
 .active {
   color: #fff;
