+++ src/components/templates/authRequiredModal/AuthRequiredModal.jsx
