+++ src/components/templates/authRequiredModal/index.js
