+++ src/components/templates/corsErrorModal/CorsErrorModal.jsx
