+++ src/components/templates/corsErrorModal/index.js
