+++ src/components/templates/index.js
